chore: moved pages inside admin routes
