+++ src/pages/admin/elements/index.tsx
@@ -1,14 +1,14 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import useDarkMode from '../../hooks/use-dark-theme'
+import useDarkMode from '../../../hooks/use-dark-theme'
 import { List } from 'layouts'
-import { IElement } from '../../models/element'
+import { IElement } from '../../../models/element'
 import { GetServerSideProps, NextPage } from 'next'
 import router from 'next/router'
 import { parseCookies } from 'nookies'
 import { useState } from 'react'
 import { FiEdit2, FiTrash } from 'react-icons/fi'
-import getAPIClient from '../../services/auth/api-ssr'
-import { deleteElement } from '../../services/elements/delete'
+import getAPIClient from '../../../services/auth/api-ssr'
+import { deleteElement } from '../../../services/elements/delete'
 import Swal from 'sweetalert2'
 
 const headers = [
@@ -87,7 +87,7 @@ const Elements: NextPage<{ data: IElement[] }> = ({ data }) => {
       headers={headers}
       onChangeSearch={(search) => setSearch(search)}
       handleSearch={handleSearchElements}
-      handleCreate={() => router.push('/elements/create')}
+      handleCreate={() => router.push('/admin/elements/create')}
       rows={elements.map((element) => ({
         ...element,
         atomic_mass: () => element.atomic_mass.toFixed(5),
@@ -134,7 +134,7 @@ export const getServerSideProps: GetServerSideProps = async (ctx) => {
   if (!token) {
     return {
       redirect: {
-        destination: '/login',
+        destination: '/',
         permanent: false
       }
     }
