+++ src/pages/admin/elements/update/[id].tsx
@@ -1,16 +1,22 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import useDarkMode from '../../../../hooks/use-dark-theme'
+import { IElement } from '../../../../models/element'
+import { FormData } from '../../../../models/elements-form'
+import { GetStaticPaths, GetStaticProps, NextPage } from 'next'
+import { getElementById } from '../../../../services/elements/get-by-id'
 import { yupResolver } from '@hookform/resolvers/yup'
-import { Button, ElementHeader, Input } from '../../components'
+import { Button, ElementHeader, Input } from '../../../../components'
 import { SubmitHandler, useForm } from 'react-hook-form'
-import { FormData } from '../../models/elements-form'
 import Link from 'next/link'
-import useDarkMode from '../../hooks/use-dark-theme'
 import router from 'next/router'
-import { createElement } from '../../services/elements/create'
+import { useCallback, useEffect, useState } from 'react'
+import { updateElement } from '../../../../services/elements/update'
 import Swal from 'sweetalert2'
 import * as yup from 'yup'
 
-const Create = () => {
+const Update: NextPage<{ elementData: IElement }> = ({ elementData }) => {
+  const { isDarkMode } = useDarkMode()
+  const [element, setElement] = useState<IElement | null>(elementData)
   const schema = yup.object({
     name: yup
       .string()
@@ -40,13 +46,11 @@ const Create = () => {
     element_img: yup.string().nullable(),
     enabled: yup.bool().default(true)
   })
-  const { isDarkMode } = useDarkMode()
-  const { register, handleSubmit, formState, reset, watch } = useForm<FormData>(
-    {
+  const { register, handleSubmit, formState, reset, setValue, watch } =
+    useForm<FormData>({
       mode: 'onChange',
       resolver: yupResolver(schema)
-    }
-  )
+    })
 
   const handleSave: SubmitHandler<FormData> = ({
     appearance,
@@ -74,7 +78,7 @@ const Create = () => {
     xpos,
     ypos
   }) => {
-    createElement({
+    updateElement({
       appearance,
       atomic_mass,
       boil,
@@ -101,8 +105,8 @@ const Create = () => {
       ypos
     })
       .then((response) => {
-        if (response.status === 201) {
-          Swal.fire('Success!', 'Element was created!', 'success')
+        if (response.status === 200) {
+          Swal.fire('Success!', 'Element was updated!', 'success')
           reset()
           router.push('/elements')
         }
@@ -110,13 +114,81 @@ const Create = () => {
       .catch((error: any) => {
         Swal.fire(
           'Error',
-          `Error to create this element! <br> ${error.response.data.message}`,
+          `Error to update this element! <br> ${error.response.data.message}`,
           'error'
         )
       })
   }
 
   const handleCancel = () => router.push('/elements')
+
+  const populateFields = useCallback(() => {
+    console.log({ element })
+    if (element !== null) {
+      const {
+        appearance,
+        atomic_mass,
+        boil,
+        category,
+        density,
+        discovered_by,
+        electron_affinity,
+        electron_configuration,
+        electron_configuration_semantic,
+        element_img,
+        enabled,
+        melt,
+        molar_heat,
+        name,
+        named_by,
+        number,
+        period,
+        phase,
+        source,
+        spectral_img,
+        summary,
+        symbol,
+        xpos,
+        ypos
+      } = element
+
+      setValue('appearance', appearance)
+      setValue('atomic_mass', atomic_mass)
+      setValue('boil', boil)
+      setValue('category', category)
+      setValue('density', density)
+
+      setValue('discovered_by', discovered_by)
+      setValue('electron_affinity', electron_affinity)
+      setValue('electron_configuration', electron_configuration)
+      setValue(
+        'electron_configuration_semantic',
+        electron_configuration_semantic
+      )
+      setValue('element_img', element_img)
+
+      setValue('melt', melt)
+      setValue('molar_heat', molar_heat)
+      setValue('name', name)
+      setValue('named_by', named_by)
+      setValue('number', number)
+
+      setValue('period', period)
+      setValue('phase', phase)
+      setValue('source', source)
+      setValue('spectral_img', spectral_img)
+      setValue('symbol', symbol)
+
+      setValue('summary', summary)
+      setValue('enabled', enabled)
+      setValue('xpos', xpos)
+      setValue('ypos', ypos)
+    }
+  }, [element, setValue])
+
+  useEffect(() => {
+    populateFields()
+  }, [populateFields])
 
   return (
     <div className="flex flex-col w-11/12 h-full items-centers">
@@ -125,19 +197,19 @@ const Create = () => {
         <h4 className="flex gap-2 items-center text-gray-600 text-sm mb-10">
           <Link href="/">Home</Link> {'>'}
           <Link href="/elements/">Elements</Link> {'>'}
-          <span className="text-blue-500 ">Create Element</span>
+          <span className="text-blue-500 ">Update Element</span>
         </h4>
         <h1
           className={`${
             isDarkMode ? 'text-gray-100' : 'text-gray-900'
           } text-left text-3xl`}
         >
-          Create Element
+          Update Element
         </h1>
       </div>
 
       {/* MAIN CONTAINER */}
-      <div className="flex flex-col mt-5">
+      <div className="flex flex-col">
         <div className="-my-2 sm:-mx-6 lg:-mx-8">
           <div className="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
             <div
@@ -147,15 +219,17 @@ const Create = () => {
             >
               <div className="flex flex-col gap-10 w-full p-10">
                 <div>
-                  <ElementHeader
-                    category={watch('category') || 'nobleGases'}
-                    atomic_mass={watch('atomic_mass')}
-                    element_img={watch('element_img') || ''}
-                    name={watch('name') || 'Unknown'}
-                    number={watch('number') || 0}
-                    symbol={watch('symbol') || 'Unk'}
-                    height="300px"
-                  />
+                  {element && (
+                    <ElementHeader
+                      category={watch('category') || element.category}
+                      atomic_mass={watch('atomic_mass') || element.atomic_mass}
+                      element_img={watch('element_img') || element.element_img}
+                      name={watch('name') || element.name}
+                      number={watch('number') || element.number}
+                      symbol={watch('symbol') || element.symbol}
+                      height="300px"
+                    />
+                  )}
                 </div>
 
                 <div className="flex gap-10 w-full">
@@ -334,7 +408,7 @@ const Create = () => {
               <div
                 className={`${
                   isDarkMode ? 'bg-gray-900' : 'bg-gray-50'
-                } px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse`}
+                } px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse rounded sm:rounded-lg`}
               >
                 <Button
                   type="button"
@@ -357,4 +431,26 @@ const Create = () => {
   )
 }
 
-export default Create
+export default Update
+
+export const getStaticPaths: GetStaticPaths = async () => {
+  let size = Array.from(Array(119).keys())
+  size = size.map((size) => size + 1)
+
+  const paths = size.map((id) => ({
+    params: { id: id.toString() }
+  }))
+
+  return { paths, fallback: 'blocking' }
+}
+
+export const getStaticProps: GetStaticProps = async ({ params }) => {
+  const id = params?.id
+
+  const element = await getElementById({ id: id as string })
+
+  return {
+    props: { elementData: element.data },
+    revalidate: 1
+  }
+}
